+++ src/main/java/dungeon/Test.java
@@ -1,0 +1,49 @@
+package dungeon;
+
+import javafx.application.Application;
+import javafx.scene.Group;
+import javafx.scene.Scene;
+import javafx.scene.paint.Color;
+import javafx.scene.shape.Rectangle;
+import javafx.stage.Stage;
+
+public class Test extends Application {
+    private static final int NUMBER_OF_BLOCKS = 10;
+    private static final int BLOCKSIZE = 40;
+    
+    private static final int SPACING = 2;
+    private static final int WIDTH = (SPACING + BLOCKSIZE) * NUMBER_OF_BLOCKS;
+    private static final int HEIGHT = (SPACING + BLOCKSIZE) * NUMBER_OF_BLOCKS;
+
+    public static void main(String[] args) {
+        Application.launch(args);
+    }
+
+    @Override
+    public void start(Stage primaryStage) throws Exception {
+        final Group group = new Group();
+        for (int iy = 0; iy < NUMBER_OF_BLOCKS; iy++) {
+            for (int ix = 0; ix < NUMBER_OF_BLOCKS; ix++) {
+
+                final Rectangle rectangle = new Rectangle(BLOCKSIZE, BLOCKSIZE);
+                rectangle.setTranslateX(ix*(BLOCKSIZE + SPACING));
+                rectangle.setTranslateY(iy*(SPACING + BLOCKSIZE));
+                rectangle.setFill(Color.GREEN);
+                rectangle.setOnMouseEntered(e -> {
+                    rectangle.setStroke(Color.DARKGRAY);
+                    rectangle.setStrokeWidth(3);
+                });
+                rectangle.setOnMouseClicked(e -> {
+                    rectangle.setFill(Color.BROWN);
+                });
+                rectangle.setOnMouseExited(e -> rectangle.setStrokeWidth(0));
+                group.getChildren().add(rectangle);
+            }
+        }
+
+        final Scene scene = new Scene(group, WIDTH, HEIGHT);
+        
+        primaryStage.setScene(scene);
+        primaryStage.show();
+    }
+}
